chore: improve hex maps
