+++ packages/common/src/lib/hexagonal-map.ts
@@ -1,0 +1,8 @@
+import { Cell } from './cell';
+
+export interface HexagonalMap {
+  readonly id: string;
+  readonly map: Cell[][];
+  readonly maxX: number;
+  readonly maxY: number;
+}
