+++ packages/common/src/lib/outdoor.ts
@@ -1,7 +1,7 @@
-import { HexMap } from './hex-map';
+import { HexagonalMap } from './hexagonal-map';
 import { GeoCoordinates } from './types';
 
-export interface Outdoor extends HexMap {
+export interface Outdoor extends HexagonalMap {
   readonly terrainType: string;
   readonly boundaries: [GeoCoordinates, GeoCoordinates];
 }
